Рефакторинг маршрутизации и состояния приложения, добавление нового файла конфигурации маршрутизатора и упрощение управления состоянием
