+++ src/App.tsx
@@ -1,38 +1,13 @@
-import { checkAuth, getCourse, getLessons, validEmail } from "@/api";
+import { configureRouter } from "@/routerConfig";
 import CourseMenu from "@/pages/CourseMenu";
 import Navbar from "@/modules/Navbar";
 import LessonPage from "@/pages/Lesson";
-import NotFoundView from "@/nonFound";
 import Settings from "@/pages/Settings";
 import MainMenu from "@/pages/MainMenuContent";
-import { defineStore } from "@/ourReact/jsx-runtime";
-import { router } from "@/router";
-import { UserProfile } from "@/types/users";
-import { CourseOpen } from "@/types/courseMenu";
 import WindowLogin from "@/modules/WindowLogin";
-import WindowALert from "./components/WindowALert/WindowALert";
-import { Toast } from "@/types/notifications";
-import removeToast from "./components/WindowALert/logic/remove";
-export const [useCourseOpen, setCourseOpen] = defineStore(
-  "CourseOpen",
-  {} as CourseOpen
-);
-export const [useLessonID, setLessonID] = defineStore<number | false>(
-  "LessonData",
-  false
-);
-export const [usePage, setPage] = defineStore("Page", "MainMenu");
-export const [useMenu, setMenu] = defineStore("menu", false);
-export const [useUser, setUser] = defineStore("auth", false as UserProfile);
-export const [useLoginWindow, setLoginWindow] = defineStore(
-  "loginWindow",
-  false
-);
-export const [useToast, setToast] = defineStore("toast1", {
-  data: [] as Toast,
-  count: 0 as number,
-});
-export const [isSearch, setSearch] = defineStore("search", "");
+import ToastContainer from "./modules/ToastContainer";
+
+import { usePage, useLoginWindow } from "@/stores";
 
 // if ("serviceWorker" in navigator) {
 //   navigator.serviceWorker
@@ -45,93 +20,7 @@ export const [isSearch, setSearch] = defineStore("search", "");
 //     });
 // }
 
-router.register("/", "MainMenu", () => {
-  console.log("Главная страница");
-});
-
-router.register("/settings", "Setting", () => {
-  console.log("Настройки");
-});
-
-router.register("/course/{id}", "CourseMenu", async () => {
-  const match = router.matchPathToRoute(location.pathname);
-  const id = Number(match?.params.id);
-
-  if (!isNaN(id)) {
-    const course = await getCourse(id);
-    if (course) {
-      setCourseOpen(course);
-    } else {
-      // TODO: Выводить 404 если не существует курса
-      console.error("курс не найден");
-      setCourseOpen({});
-    }
-  }
-});
-
-router.register("/course/{id}/lessons", "Lessons", async () => {
-  const matched = router.matchPathToRoute(location.pathname);
-  const id = Number(matched?.params.id);
-
-  console.log("Урок");
-
-  if (!isNaN(id)) {
-    const course = await getCourse(id);
-    const courseId = course.id;
-    if (courseId === undefined) {
-      console.error("Ошибка");
-      return;
-    }
-    const lessonId = await getLessons(courseId);
-    if (lessonId === undefined) {
-      console.error("Ошибка");
-      return;
-    }
-    if (course) {
-      setCourseOpen(course);
-      setLessonID(lessonId.lesson.header.Points[0].lesson_id);
-    } else {
-      // TODO: Выводить 404 если не существует курса
-      console.error("курс не найден");
-      setCourseOpen({});
-    }
-  }
-});
-
-router.register("/validate/{token}", "validDate", async () => {
-  const matched = router.matchPathToRoute(location.pathname);
-  const token = matched?.params.token;
-
-  console.log("Валидация с почты");
-
-  if (token !== undefined) {
-    const valid = await validEmail(token);
-    if (valid) {
-      router.goByState("MainMenu");
-      checkAuth().then((data) => {
-        if (data === false) {
-          return;
-        }
-        console.log(data);
-        setUser({
-          name: data.name,
-          email: data.email,
-          bio: data.bio,
-          avatar_src: data.avatar_src,
-          hide_email: data.hide_email,
-        });
-      });
-    } else {
-      // TODO: Выводить 404 если не существует курса
-      console.error("Ошибка: Авторизация не удалась");
-      alert("Ошибка: Авторизация не удалась");
-      router.goByState("MainMenu");
-    }
-  }
-});
-
-router.setNotFoundView(NotFoundView);
-router.start();
+configureRouter();
 
 const App = () => {
   let content;
@@ -169,17 +58,7 @@ const App = () => {
       <Navbar key="MainHeader" />
       {content}
       {useLoginWindow() ? <WindowLogin key="WindowLogin" /> : ""}
-      <div class="alert-admit">
-        {useToast().data.map((notify, ind) => (
-          <WindowALert
-            key={`windowAlert-${ind}`}
-            type={notify.type}
-            message={notify.message}
-            disappear={notify.disappear}
-            onClose={() => removeToast(notify.id)}
-          />
-        ))}
-      </div>
+      <ToastContainer key="ToastContainer" />
     </div>
   );
 };
